+++ core/__init__.py
@@ -1,6 +1,6 @@
 """核心模块"""
 
-APP_VERSION = "1.5.6.0"
+APP_VERSION = "1.5.6.7"
 
 from .data_models import (
     ShortcutType, ShortcutItem, Folder, AppSettings, AppData,